crate: move BootInformation to dedicated module
This improves code quality
